VideoBackends/Software: Replace memset in SWVertexLoader's DrawCurrentBatch()
In 8560eecd494482146d289a06d8bd4c6033a6fa14, this memset was changed to
cast m_vertex to void* in order to silence class-memaccess warnings.
While this works, this isn't necessarily a proper way of fixing the
underlying issue -- the fact that a member in the InputVertexData struct
is non-trivial. It's a requirement for the output pointer of std::memset
to point to an object that is trivially copyable or else the behavior is
undefined.

This is not the case here, so this is technically invoking undefined
behavior (though in practice it will still work, but that's not the
point). Prior to the amendments to Vec3 in the previous commit, Vec3's
default constructor technically didn't initialize its members, which
this std::memset call was overriding essentially.

Now that we properly initialize the data members of Vec3 in the default
constructor, we can simply convert this over to using aggregate
initialization, providing the same behavior, but without the potential
to invoke undefined behavior.