InputCommon/ExpressionParser: Make ControlQualifier's members private
Rather than allow freeform modification of the class' internals, we can
make these private and build the respective API around the setting and
getting. In particular, setting the device qualifier signifies the
presence of a device, so this is better placed in its own function so
this isn't potentially forgotten if the relevant code is moved around or
refactored in the future.